scoring function improvement - backfire effect
The numeric magnitude that represents the backfire-effect in the scoring function has been doubled so that the backfire effect is twice as avoided than previous scoring function. This should better supress concepts from being shown to users that trigger the backfire effect (better suppresses concepts in feed that demotivates the user).